lnm: Remove unnecessary adapters in favor of require/opaque
This gets rid of a double-contract boundary problem that was
found. The math module's contract could not be verified so was
left in and the module turned into no-check. However, this was
requiring math/statistics, a typed library, inducing a double
boundary only in the optimized version. This change fixes that
(although we still have an adapter for fake-plot, I couldn't
get rid of this one because of the reliance on an opaque Pict
type that needs to remain opaque).